+++ src/vue/modules/assets/balance-explorer/store/index.js
@@ -46,10 +46,12 @@ export const actions = {
 
 export const getters = {
   [types.accountId]: state => state.accountId,
-  [types.assets]: state => state.assets.map(asset => {
-    const balance = state.balances.find(b => b.asset.id === asset.id) || ''
-    return new Asset(asset, balance ? balance.state.available : '')
-  }),
+  [types.assets]: state => state.assets
+    .map(asset => {
+      const balance = state.balances.find(b => b.asset.id === asset.id)
+      return new Asset(asset, balance ? balance.state.available : '')
+    })
+    .sort((a, b) => b.balance - a.balance),
   [types.kycRequiredAssetType]: state => state.kycRequiredAssetType,
 }